Keep launcher icon + splash purple; fix VoiceChanger float/double min error
- Revert launcher assets (all foregrounds, legacy launchers, adaptive
  backgrounds) and Android 12+ splash back to the original violet-purple
  brand; in-app UI stays light green (accents, intro badges)
- VoiceChanger.cpp: cast lo/hi to double so std::min/std::max deduce

+++ TMessagesProj/jni/voip/tgcalls/VoiceChanger.cpp
@@ -51,7 +51,7 @@ const VoicePreset *VoiceChanger::PresetById(int id) {
 }
 
 static float ClampFloat(double v, float lo, float hi) {
-    return (float)std::max(lo, std::min(hi, v));
+    return (float)std::max((double)lo, std::min((double)hi, v));
 }
 
 static double OnePoleCoefficient(double cutoffHz, int sampleRate) {
